#145 refract the cts structure to bypass the class duplication
